Removed Event Processor Base and moved its contents to the concrete class
